module: move source hooks into their own files
