+++ lib/internal/modules/esm/loader.js
@@ -20,9 +20,13 @@ const ModuleJob = require('internal/modules/esm/module_job');
 
 const { defaultResolve } = require('internal/modules/esm/resolve');
 const { defaultGetFormat } = require('internal/modules/esm/get_format');
+const { defaultGetSource } = require(
+  'internal/modules/esm/get_source');
+const { defaultTransformSource } = require(
+  'internal/modules/esm/transform_source');
 const createDynamicModule = require(
   'internal/modules/esm/create_dynamic_module');
-const { defaultGetSource, defaultTransformSource, translators } = require(
+const { translators } = require(
   'internal/modules/esm/translators');
 const { getOptionValue } = require('internal/options');
