+++ lib/internal/modules/esm/transform_source.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function defaultTransformSource(source, { url, format } = {},
+                                defaultTransformSource) {
+  return { source };
+}
+exports.defaultTransformSource = defaultTransformSource;
