+++ lib/internal/modules/esm/translators.js
@@ -9,26 +9,22 @@ const {
   StringPrototypeReplace,
 } = primordials;
 
-const { Buffer } = require('buffer');
-
 const {
   stripBOM,
   loadNativeModule
 } = require('internal/modules/cjs/helpers');
 const CJSModule = require('internal/modules/cjs/loader').Module;
 const internalURLModule = require('internal/url');
+const { defaultGetSource } = require(
+  'internal/modules/esm/get_source');
+const { defaultTransformSource } = require(
+  'internal/modules/esm/transform_source');
 const createDynamicModule = require(
   'internal/modules/esm/create_dynamic_module');
-const fs = require('fs');
 const { fileURLToPath, URL } = require('url');
 const { debuglog } = require('internal/util/debuglog');
-const { promisify, emitExperimentalWarning } = require('internal/util');
-const {
-  ERR_INVALID_URL,
-  ERR_INVALID_URL_SCHEME,
-  ERR_UNKNOWN_BUILTIN_MODULE
-} = require('internal/errors').codes;
-const readFileAsync = promisify(fs.readFile);
+const { emitExperimentalWarning } = require('internal/util');
+const { ERR_UNKNOWN_BUILTIN_MODULE } = require('internal/errors').codes;
 const { maybeCacheSourceMap } = require('internal/source_map/source_map_cache');
 const moduleWrap = internalBinding('module_wrap');
 const { ModuleWrap } = moduleWrap;
@@ -37,35 +33,6 @@ const debug = debuglog('esm');
 
 const translators = new SafeMap();
 exports.translators = translators;
-
-const DATA_URL_PATTERN = /^[^/]+\/[^,;]+(?:[^,]*?)(;base64)?,([\s\S]*)$/;
-
-async function defaultGetSource(url, { format } = {}, defaultGetSource) {
-  const parsed = new URL(url);
-  if (parsed.protocol === 'file:') {
-    return {
-      source: await readFileAsync(parsed)
-    };
-  } else if (parsed.protocol === 'data:') {
-    const match = DATA_URL_PATTERN.exec(parsed.pathname);
-    if (!match) {
-      throw new ERR_INVALID_URL(url);
-    }
-    const [ , base64, body ] = match;
-    return {
-      source: Buffer.from(body, base64 ? 'base64' : 'utf8')
-    };
-  } else {
-    throw new ERR_INVALID_URL_SCHEME(['file', 'data']);
-  }
-}
-exports.defaultGetSource = defaultGetSource;
-
-function defaultTransformSource(source, { url, format } = {},
-                                defaultTransformSource) {
-  return { source };
-}
-exports.defaultTransformSource = defaultTransformSource;
 
 function errPath(url) {
   const parsed = new URL(url);
