+++ test/parallel/test-bootstrap-modules.js
@@ -51,10 +51,12 @@ const expectedModules = new Set([
   'NativeModule internal/modules/cjs/loader',
   'NativeModule internal/modules/esm/create_dynamic_module',
   'NativeModule internal/modules/esm/get_format',
+  'NativeModule internal/modules/esm/get_source',
   'NativeModule internal/modules/esm/loader',
   'NativeModule internal/modules/esm/module_job',
   'NativeModule internal/modules/esm/module_map',
   'NativeModule internal/modules/esm/resolve',
+  'NativeModule internal/modules/esm/transform_source',
   'NativeModule internal/modules/esm/translators',
   'NativeModule internal/process/esm_loader',
   'NativeModule internal/options',
